Rename Styles to StylesMap.
